bridgyv2-main/pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "faiss-cpu>=1.10.0",
    "langchain-community>=0.3.18",
    "langchain>=0.3.19",
    "openai>=1.65.2",
    "python-dotenv>=1.0.1",
    "streamlit>=1.42.2",
    "tiktoken>=0.9.0",
    "intersight>=1.0.11",
    "trafilatura>=2.0.0",
    "langchain-openai>=0.3.7",
    "pillow>=11.1.0",
    "pypdf>=5.3.1",
]
